requirements.txt
boto3
ultralytics
python_dotenv